Since the n-m titration calculator is complete, the functions in these files have been moved to the main titration_module.py
